gsc-monitoring/pyproject.toml
[project]
name = "gsc-issues"
version = "0.1.0"
description = "Google Search Console issues fetcher"
requires-python = ">=3.10"
dependencies = [
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.152.0",
]

[project.scripts]
gsc-fetch = "gsc_fetch_issues:main"
